Start refactoring convert_n_test()
This subroutine is used in Pod-Html's test suite.  Currently, it takes a
list of arguments in which:

        1: Root of name of test file in t/
        2: Test description displayed by harness
        3: List of command-line switches for 'pod2html',
           some of which may override default values

When debugging a function, it is useful to be able to dump the
function's arguments before we call the function, and to be able to do
so confidently that what we see in the dump is what we pass to the
function.

However, switches in the third position may silently override default
switches defined by convert_n_test.  That can make debugging a bit
murky.  Let's start by pulling the default switches into a lookup table.

For: GH 16140