Parenthesized variable declaration must follow 'catch'
Experimentation showed that, following the 'catch' keyword, omitting
either:
* the variable from within the parentheses; or
* both the parentheses and the variable
caused a syntax error.  Add one test to demonstrate that we can capture
the syntax error.  Update the documentation to insist upon the variable
declaration.

Note:  While we capture the syntax error here, the content of the error
is sub-optimal as it does not direct the user precisely to the source of
the error.