my_strftime(): Don't assume empty %p is an error
The strftime() format %p yields an am/pm indicator.  Many locales don't
have this distinction, and so a format containing just this specifier
will return an empty string.  Prior to this commit, that was considered
an error.  What was going on is we kept enlarging the buffer to find a
non-empty return, to no avail, and so treat it as an error.  This commit
special cases this to not make it an error.  (It's hard to imagine that
the am/pm string alone would exceed the 200 byte limit.)