Also set RXf_RTRIM if the pattern is /\s*$/ (without the //u flag).
Without the //u flag, \s and [[:space:]] are compiled to POSIXD ops.
The POSIXD op behaves differently depending on whether the target string has
SVf_UTF8 set.

One could in theory continue and handle the case of the //a flag (POSIXA
ops), but this doesn't seem worth it for an optimisation, as it is unlikely
to be common on regexs that are intended to remove "generic" whitespace from
the end of a string.