Untangle the logic for determining when "DYNAMIC_ENV_FETCH" will be called.
This does now mean that we call mg_find() on all hashes before the fetch,
whereas previously we only called it if the (initial) fetch failed, but it
lets us distinguish all the cases where what is requested as an LVALUE fetch
will actually recurse into hv_common() or return early, and hence not reach
the store code.

Also split out the decision about "will_recurse_because_magical" and code
"in_collision" as a direct lookup on HvARRAY, which decouples it from the
code related to linking the newly allocated HE into the hash.