XXX Add a key_mask to Perl_ABH_Table, defaulting to HVhek_WASUTF8.
This is the user-visible hash behaviour - keys that could be represented as
UTF-8 or Octets are considered equal, but we need to remember which form
they were stored in.

However, the shared string table needs to treat them as distinct, and as it
shares the code we have to support both variants.

Also make the size default to sizeof(HE *), and automatically build the
hashtable if we are called to insert on a NULL table.