XXX ithreads cloning for ABH hashes.
FIXME - cloning hashes with unshared keys not implemented (yet)
FIXME - this approach can't clone the hash iterator

To clone the hash iteartor we'd need to take the same approach as
Perl_ABH_fast_HV_copy. But then we get into a struggle with static
functions - those that want to remain in hv_abh.c, and
S_share_hek_flags().