Change S_new_body to static inline, and use it instead of a macro.
The macro new_body_inline() existed to provide an inline implementation for
hot code paths, with a small function for the others. Now that we have
proper inline functions, change the function to be inlined, and use it
instead of the macro. There is no change to the code generated for the
hot paths. This makes the source simpler.

Yes, this will also inline the function in a couple of other places, but
these days optimising compilers often do that anyway for small functions.
Even if they aren't for this function (as was), the size of the function
itself is comparable with the calling overheads, so it's not a massive
penalty.

Retain the macro new_body_inline, and split the parameters to the macro.
This will make the next commit simpler.