Add HvIS_EMPTY(), returning true if a hash contains no keys.
This is intended as an efficient test to decide whether to skip iterating
the hash (and any setup needed beforehand). It may return false even if the
hash is actually empty, if this can't trivially be determined, such as for
tied hashes and hashes which only contain placeholders.

Use this in preference to a truth check on HvARRAY to determine whether to
skip iteration, as it will detect more "empty" hashes.